refactor: split monolithic files into modular components
