+++ src/styles/admin.css
@@ -1,0 +1,84 @@
+.admin-tabs {
+  display: flex;
+  gap: 4px;
+  flex-wrap: wrap;
+  margin-bottom: 1.5rem;
+  background: var(--paper);
+  border: 1px solid var(--line);
+  border-radius: 8px;
+  padding: 6px;
+  box-shadow: var(--shadow);
+}
+
+.admin-tab {
+  display: inline-flex;
+  align-items: center;
+  gap: 6px;
+  padding: 8px 14px;
+  border: none;
+  border-radius: 6px;
+  background: transparent;
+  color: var(--muted);
+  font-size: 0.88rem;
+  font-weight: 500;
+  cursor: pointer;
+  transition: background 0.15s, color 0.15s;
+}
+
+.admin-tab:hover {
+  background: var(--wash);
+  color: var(--ink);
+}
+
+.admin-tab.active {
+  background: var(--blue);
+  color: #fff;
+}
+
+.admin-section {
+  animation: fadeIn 0.2s ease;
+}
+
+.admin-page .metrics-grid {
+  grid-template-columns: repeat(4, minmax(0, 1fr));
+}
+
+@media (max-width: 980px) {
+  .admin-page .metrics-grid {
+    grid-template-columns: repeat(2, minmax(0, 1fr));
+  }
+}
+
+@media (max-width: 680px) {
+  .admin-page .metrics-grid {
+    grid-template-columns: repeat(2, minmax(0, 1fr));
+  }
+}
+
+.portfolio-manager-grid {
+  display: grid;
+  grid-template-columns: repeat(3, minmax(0, 1fr));
+  gap: 14px;
+}
+
+.portfolio-manager-card {
+  border: 1px solid var(--line);
+  border-radius: 8px;
+  background: var(--paper);
+  overflow: hidden;
+}
+
+.portfolio-manager-card img {
+  aspect-ratio: 3 / 2;
+  object-fit: cover;
+  background: #d7dee4;
+}
+
+.portfolio-manager-card .card-body {
+  padding: 12px;
+}
+
+.portfolio-manager-card h3 {
+  font-size: 0.95rem;
+  margin-bottom: 4px;
+}
