+++ src/styles/base.css
@@ -1,0 +1,77 @@
+:root {
+  color-scheme: light;
+  --ink: #151515;
+  --muted: #68707a;
+  --line: #dde3e7;
+  --paper: #ffffff;
+  --wash: #f3f6f8;
+  --charcoal: #111820;
+  --blue: #2364aa;
+  --blue-soft: #e6f0fb;
+  --green: #1f8a5b;
+  --green-soft: #e6f6ee;
+  --red: #bd3a3a;
+  --red-soft: #fdeaea;
+  --amber: #b36a00;
+  --amber-soft: #fff3d6;
+  --shadow: 0 16px 40px rgba(18, 31, 45, 0.08);
+  font-family: Inter, ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif;
+}
+
+* {
+  box-sizing: border-box;
+}
+
+body {
+  margin: 0;
+  background: var(--wash);
+  color: var(--ink);
+}
+
+a {
+  color: inherit;
+  text-decoration: none;
+}
+
+button,
+input,
+select,
+textarea {
+  font: inherit;
+}
+
+img {
+  display: block;
+  max-width: 100%;
+}
+
+h1,
+h2,
+h3,
+p {
+  margin-top: 0;
+}
+
+h1 {
+  font-size: clamp(34px, 5vw, 58px);
+  line-height: 1;
+  margin-bottom: 14px;
+}
+
+h2 {
+  font-size: clamp(24px, 3vw, 36px);
+  line-height: 1.05;
+  margin-bottom: 18px;
+}
+
+h3 {
+  margin-bottom: 8px;
+}
+
+.muted,
+.feature p,
+.soft-row p,
+.table-row p,
+.mini-note {
+  color: var(--muted);
+}
