+++ src/styles/cards.css
@@ -1,0 +1,80 @@
+.feature,
+.panel,
+.photographer-card,
+.metric,
+.auth-card {
+  border: 1px solid var(--line);
+  border-radius: 8px;
+  background: var(--paper);
+  box-shadow: var(--shadow);
+}
+
+.feature {
+  padding: 22px;
+}
+
+.feature span,
+.metric span {
+  display: grid;
+  place-items: center;
+  width: 42px;
+  height: 42px;
+  border-radius: 8px;
+  color: var(--blue);
+  background: var(--blue-soft);
+  margin-bottom: 18px;
+}
+
+.photographer-card {
+  overflow: hidden;
+}
+
+.card-image {
+  position: relative;
+  aspect-ratio: 4 / 3;
+  background: #c8d4de;
+}
+
+.card-image img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.card-body,
+.panel,
+.auth-card {
+  padding: 18px;
+}
+
+.card-title-row,
+.card-footer,
+.table-row {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 12px;
+}
+
+.card-title-row h3 {
+  margin-bottom: 3px;
+}
+
+.card-title-row p,
+.card-footer {
+  margin: 0;
+}
+
+.card-title-row p,
+.rating,
+.price,
+.hero-meta span {
+  display: inline-flex;
+  align-items: center;
+  gap: 5px;
+}
+
+.rating {
+  color: #946200;
+  font-weight: 800;
+}
