+++ src/styles/forms.css
@@ -1,0 +1,84 @@
+label {
+  display: grid;
+  gap: 7px;
+  color: var(--muted);
+  font-size: 13px;
+  font-weight: 700;
+}
+
+input,
+select,
+textarea {
+  width: 100%;
+  border: 1px solid var(--line);
+  border-radius: 8px;
+  background: var(--paper);
+  color: var(--ink);
+  padding: 11px 12px;
+  min-height: 42px;
+  outline: none;
+}
+
+textarea {
+  min-height: 96px;
+  resize: vertical;
+}
+
+input:focus,
+select:focus,
+textarea:focus {
+  border-color: var(--blue);
+  box-shadow: 0 0 0 3px var(--blue-soft);
+}
+
+.segmented {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  padding: 4px;
+  border-radius: 8px;
+  background: var(--wash);
+}
+
+.segmented button {
+  border: 0;
+  border-radius: 8px;
+  min-height: 38px;
+  background: transparent;
+  cursor: pointer;
+  font-weight: 800;
+}
+
+.segmented .active {
+  background: var(--paper);
+  box-shadow: 0 3px 16px rgba(18, 31, 45, 0.08);
+}
+
+.form-error,
+.form-success {
+  border-radius: 8px;
+  padding: 10px 12px;
+  margin: 0;
+  font-weight: 700;
+}
+
+.form-error {
+  color: var(--red);
+  background: var(--red-soft);
+}
+
+.form-success {
+  color: var(--green);
+  background: var(--green-soft);
+}
+
+.inline-form {
+  display: grid;
+  grid-template-columns: 1fr 1.3fr 120px auto;
+  gap: 10px;
+}
+
+.form-row {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 10px;
+}
