+++ src/styles/index.css
@@ -1,0 +1,17 @@
+@import "./base.css";
+@import "./layout.css";
+@import "./buttons.css";
+@import "./forms.css";
+@import "./cards.css";
+@import "./tags-badges.css";
+@import "./grid.css";
+@import "./profile.css";
+@import "./booking.css";
+@import "./messages.css";
+@import "./modals.css";
+@import "./toasts.css";
+@import "./admin.css";
+@import "./tabs.css";
+@import "./tables.css";
+@import "./timeline.css";
+@import "./responsive.css";
