+++ src/styles/layout.css
@@ -1,0 +1,194 @@
+.app-shell {
+  min-height: 100vh;
+}
+
+.site-header {
+  position: sticky;
+  top: 0;
+  z-index: 20;
+  display: grid;
+  grid-template-columns: auto 1fr auto;
+  align-items: center;
+  gap: 20px;
+  padding: 14px clamp(18px, 4vw, 56px);
+  background: rgba(255, 255, 255, 0.88);
+  border-bottom: 1px solid var(--line);
+  backdrop-filter: blur(16px);
+}
+
+.brand,
+.site-nav,
+.header-actions,
+.inline-actions {
+  display: flex;
+  align-items: center;
+}
+
+.brand {
+  gap: 10px;
+  font-weight: 800;
+  letter-spacing: 0;
+}
+
+.brand-mark {
+  display: grid;
+  place-items: center;
+  width: 34px;
+  height: 34px;
+  border-radius: 8px;
+  color: #fff;
+  background: var(--charcoal);
+}
+
+.brand-mark.large {
+  width: 46px;
+  height: 46px;
+}
+
+.site-nav {
+  justify-content: center;
+  gap: 8px;
+}
+
+.page {
+  width: min(1180px, calc(100% - 32px));
+  margin: 0 auto;
+  padding: 42px 0;
+}
+
+.page.narrow {
+  width: min(720px, calc(100% - 32px));
+}
+
+.hero {
+  min-height: calc(100vh - 64px);
+  position: relative;
+  display: grid;
+  align-items: end;
+  overflow: hidden;
+  background: var(--charcoal);
+}
+
+.hero img {
+  position: absolute;
+  inset: 0;
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  opacity: 0.72;
+}
+
+.hero::after {
+  content: "";
+  position: absolute;
+  inset: 0;
+  background: linear-gradient(90deg, rgba(10, 14, 20, 0.88), rgba(10, 14, 20, 0.42) 58%, rgba(10, 14, 20, 0.18));
+}
+
+.hero-content {
+  position: relative;
+  z-index: 1;
+  width: min(850px, calc(100% - 32px));
+  margin: 0 auto;
+  padding: 0 0 64px;
+  color: #fff;
+}
+
+.hero-content h1 {
+  margin: 14px 0 18px;
+  font-size: clamp(42px, 8vw, 86px);
+  line-height: 0.94;
+  max-width: 820px;
+}
+
+.hero-content p {
+  color: rgba(255, 255, 255, 0.82);
+  max-width: 660px;
+  font-size: 18px;
+  line-height: 1.6;
+}
+
+.eyebrow {
+  text-transform: uppercase;
+  letter-spacing: 0.12em;
+  font-size: 12px;
+  font-weight: 800;
+  color: var(--blue);
+}
+
+.hero .eyebrow {
+  color: #9ed0ff;
+}
+
+.search-band,
+.filter-bar {
+  display: grid;
+  align-items: end;
+  gap: 12px;
+  margin-top: 28px;
+  padding: 12px;
+  border-radius: 8px;
+  background: rgba(255, 255, 255, 0.94);
+  color: var(--ink);
+  box-shadow: var(--shadow);
+}
+
+.search-band {
+  grid-template-columns: 1fr auto;
+  max-width: 680px;
+}
+
+.filter-bar {
+  grid-template-columns: repeat(4, minmax(120px, 1fr)) auto;
+}
+
+.section-grid {
+  display: grid;
+  grid-template-columns: 0.9fr 1.5fr;
+  gap: 36px;
+  align-items: start;
+}
+
+.section-heading {
+  position: sticky;
+  top: 94px;
+}
+
+.feature-grid {
+  display: grid;
+  grid-template-columns: repeat(2, minmax(0, 1fr));
+  gap: 14px;
+}
+
+.split-heading {
+  display: flex;
+  align-items: end;
+  justify-content: space-between;
+  gap: 20px;
+  margin-bottom: 24px;
+}
+
+.compact-heading {
+  align-items: start;
+  margin-bottom: 12px;
+}
+
+.clamp {
+  display: -webkit-box;
+  overflow: hidden;
+  -webkit-line-clamp: 3;
+  -webkit-box-orient: vertical;
+}
+
+.empty-state {
+  border: 1px dashed var(--line);
+  border-radius: 8px;
+  padding: 48px;
+  text-align: center;
+  color: var(--muted);
+  background: var(--paper);
+}
+
+.stretch {
+  align-items: stretch;
+}
