+++ src/styles/toasts.css
@@ -1,0 +1,57 @@
+.toast-container {
+  position: fixed;
+  bottom: 24px;
+  right: 24px;
+  z-index: 100;
+  display: grid;
+  gap: 8px;
+  max-width: 400px;
+}
+
+.toast {
+  display: grid;
+  grid-template-columns: auto 1fr auto;
+  align-items: center;
+  gap: 10px;
+  padding: 12px 16px;
+  border-radius: 8px;
+  border: 1px solid var(--line);
+  background: var(--paper);
+  box-shadow: 0 8px 32px rgba(18, 31, 45, 0.14);
+  animation: toastIn 0.25s ease;
+  font-size: 0.92rem;
+}
+
+.toast-success { border-color: var(--green); color: var(--green); background: var(--green-soft); }
+.toast-error   { border-color: var(--red); color: var(--red); background: var(--red-soft); }
+.toast-warning { border-color: var(--amber); color: var(--amber); background: var(--amber-soft); }
+.toast-info    { border-color: var(--blue); color: var(--blue); background: var(--blue-soft); }
+
+.toast-icon {
+  display: grid;
+  place-items: center;
+}
+
+.toast-message {
+  line-height: 1.4;
+}
+
+.toast-close {
+  display: grid;
+  place-items: center;
+  width: 24px;
+  height: 24px;
+  border: none;
+  border-radius: 4px;
+  background: transparent;
+  cursor: pointer;
+  color: inherit;
+  opacity: 0.55;
+}
+
+.toast-close:hover { opacity: 1; }
+
+@keyframes toastIn {
+  from { opacity: 0; transform: translateY(12px); }
+  to   { opacity: 1; transform: translateY(0); }
+}
